requirement.txt
flask
tensorflow
pillow
numpy
matplotlib
